Code Complete
Renamed the files to their proper names. Added in the script for the bw_review.js. Added the initial function to add in the event listener. Added in the lightStars() function to light the stars up. Added the turnOffStars function to unlight the stars. Added the update count function to update the character counter. Added a fix so that once the user clicks on the stars it won't go away.